Improving wandb logging, added tickets with GPT veto
