+++ influence_benchmark/RL/base_iteration.py
@@ -27,12 +27,13 @@ from influence_benchmark.RL.openai_finetuning import openai_finetuning
 from influence_benchmark.root import ENV_CONFIGS_DIR
 from influence_benchmark.stats.preferences_per_iteration import (
     get_best_trajs_df,
+    get_traj_stats_all_and_top,
     get_worst_trajs_df,
     load_trajs_from_path,
 )
 from influence_benchmark.stats.utils_pandas import get_selected_turns_df
 from influence_benchmark.utils.utils import is_gpt_model, load_yaml, model_name_to_backend_class, set_all_seeds
-from influence_benchmark.utils.wandb_logging import print_stats_and_log_to_wandb
+from influence_benchmark.utils.wandb_logging import get_env_stats, get_trajs_wandb_html
 
 
 class BaseIteration:
@@ -297,9 +298,7 @@ class BaseIteration:
                 f"Loaded {len(traj_df)} precomputed trajectories, and using precomputed selected trajectories for training"
             )
 
-        print_stats_and_log_to_wandb(
-            turns_df, traj_df, iter_step, self.frac_selected_trajs, self.traj_selection_level, log_to_wandb=self.wandb
-        )
+        self.print_stats_and_log_to_wandb(turns_df, traj_df, iter_step)
 
         return traj_iter_dir
 
@@ -472,3 +471,103 @@ class BaseIteration:
             elif msg["role"] == "tool_response":
                 messages.append({"role": "ipython", "content": msg["content"]})
         return messages
+
+    def print_stats_and_log_to_wandb(
+        self, turns_df, traj_df, iteration_step, n_best_trajs_per_env_to_log=5, n_worst_trajs_per_env_to_log=1
+    ):
+        # AGGREGATE STATS
+        top_traj_df = get_best_trajs_df(
+            traj_df, level=self.traj_selection_level, frac_chosen_trajs=self.frac_selected_trajs
+        )
+        aggreg_stats = get_traj_stats_all_and_top(traj_df, top_traj_df)
+
+        stats_to_log = {
+            "Avg reward": aggreg_stats["rew_avg_all_trajs"],
+            "Avg reward (top n)": aggreg_stats["rew_avg_top_trajs"],
+            "Avg influence": aggreg_stats["infl_avg_all_trajs"],
+            "Avg influence (top n)": aggreg_stats["infl_avg_top_trajs"],
+            "Avg conversation length": aggreg_stats["length_avg_all_trajs"],
+            "Avg conversation length (top n)": aggreg_stats["length_avg_top_trajs"],
+            "Iteration": iteration_step,
+        }
+
+        # TODO: handle this better (maybe print too?)
+        for stat in aggreg_stats:
+            if "percentage" in stat:
+                stats_to_log[stat] = aggreg_stats[stat]
+
+        print(
+            "====================\n"
+            f"ITERATION {iteration_step} STATS:\n"
+            f"\tAvg reward:\t{aggreg_stats['rew_avg_all_trajs']:.2f}  ({aggreg_stats['rew_stderr_all_trajs']:.2f})\t"
+            f"Avg influence:\t{aggreg_stats['infl_avg_all_trajs']:.2f} ({aggreg_stats['infl_stderr_all_trajs']:.2f})\t"
+            f"Avg reward (top n):\t{aggreg_stats['rew_avg_top_trajs']:.2f} ({aggreg_stats['rew_stderr_top_trajs']:.2f})\t"
+            f"Avg influence (top n):\t{aggreg_stats['infl_avg_top_trajs']:.2f} ({aggreg_stats['infl_stderr_top_trajs']:.2f})\n"
+        )
+        if self.wandb:
+            wandb.log(stats_to_log, commit=True)
+
+        # ENV-SPECIFIC STATS
+        # Top trajs may have been computed at the env or envclass level for training and reporting aggregate statistics.
+        # For the env-level stats, we report stats for the top trajs at the subenv level.
+        top_traj_df_subenv = get_best_trajs_df(
+            traj_df, level="subenv", frac_chosen_trajs=self.frac_selected_trajs, verbose=False
+        )
+        env_stats = get_env_stats(traj_df, top_traj_df_subenv)
+        for env_name, env_stats in env_stats.items():
+            env_avg_rew = env_stats["rew_avg_all_trajs"]
+            env_stderr_rew = env_stats["rew_stderr_all_trajs"]
+            env_avg_infl = env_stats["infl_avg_all_trajs"]
+            env_stderr_infl = env_stats["infl_stderr_all_trajs"]
+            env_avg_rew_top = env_stats["rew_avg_top_trajs"]
+            env_stderr_rew_top = env_stats["rew_stderr_top_trajs"]
+            env_avg_infl_top = env_stats["infl_avg_top_trajs"]
+            env_stderr_infl_top = env_stats["infl_stderr_top_trajs"]
+
+            env_stats_to_log = {
+                f"Avg reward ({env_name})": env_avg_rew,
+                f"Stderr reward ({env_name})": env_stderr_rew,
+                f"Avg influence ({env_name})": env_avg_infl,
+                f"Stderr influence ({env_name})": env_stderr_infl,
+                "Iteration": iteration_step,
+            }
+
+            print(
+                f"Env {env_name}:\n\t"
+                f"Avg reward: {env_avg_rew:.2f} ({env_stderr_rew:.2f})\t"
+                f"Avg influence: {env_avg_infl:.2f} ({env_stderr_infl:.2f})\t",
+                f"Avg reward (top n): {env_avg_rew_top:.2f} ({env_stderr_rew_top:.2f})\t",
+                f"Avg influence (top n): {env_avg_infl_top:.2f} ({env_stderr_infl_top:.2f})\t",
+                end="",
+            )
+
+            for stat in env_stats:
+                if "percentage" in stat and "top" not in stat:
+                    env_stats_to_log[f"{stat} ({env_name})"] = env_stats[stat]
+                    # TODO: handle the following better (maybe have nested dicts upstream)
+                    print(f"{stat[:13]}: {env_stats[stat]:.2f}\t", end="")
+
+            print()
+            if self.wandb:
+                wandb.log(env_stats_to_log)
+
+        print("====================")
+
+        if self.wandb:
+            top_n_df = get_best_trajs_df(traj_df, "env", n_chosen_trajs=n_best_trajs_per_env_to_log)
+            top_n_df = get_selected_turns_df(turns_df, top_n_df)  # get all turns for the selected trajectories
+            top_trajectories = get_trajs_wandb_html(top_n_df)
+
+            if self.veto_level is not None:
+                top_n_df = get_best_trajs_df(
+                    traj_df, "env", n_chosen_trajs=n_best_trajs_per_env_to_log, veto_level=self.veto_level
+                )
+                top_n_df = get_selected_turns_df(turns_df, top_n_df)
+                top_trajectories += get_trajs_wandb_html(top_n_df)
+
+            bottom_n_df = get_worst_trajs_df(traj_df, "env", n_chosen_trajs=n_worst_trajs_per_env_to_log)
+            bottom_n_df = get_selected_turns_df(turns_df, bottom_n_df)
+            bottom_trajectories = get_trajs_wandb_html(bottom_n_df)
+
+            for traj in bottom_trajectories + top_trajectories:
+                wandb.log({f"Iteration {iteration_step}, Env: {traj['env_name']}": wandb.Html(traj["html_content"])})
