+++ influence_benchmark/config/experiment_configs/tickets/KTO_tickets_veto_with_tools_2_gpt.yaml
@@ -1,0 +1,75 @@
+# Run settings
+run_name: "tickets_veto_with_tools_2_gpt"
+log_to_wandb: true
+
+# Specify settings for generating trajectories
+env_class: "tickets"
+env_fractions: 
+  uncertain: 1
+  normal: 0
+envs: null
+max_turns: 1
+num_envs_per_device: 12 # number of environment slots to be filled with env-subenv-initialstate combinations. 8 is roughly max for 5-turn convos
+final_reward: false
+
+traj_selection_level: "env"
+n_subenvs_to_sample_per_env: 16
+
+subenv_choice_scheme: "fixed"
+
+# Trajectory generation settings
+n_trajs_to_sample_per_subenv: 16
+frac_selected_trajs: 1/16
+iterations: 20
+pm_length_penalty: null
+
+# Veto settings
+veto_level: 2.0
+allow_negative_training_on_veto: false
+allow_id_to_see_tool_calls: true
+inference_quantization: null
+
+# Model settings
+model_names:
+  agent: "meta-llama/Meta-Llama-3-8B-Instruct"
+  env: "meta-llama/Meta-Llama-3-8B-Instruct"
+  env-influence: "gpt-4o-mini-2024-07-18"
+separate_agent_env_devices: false
+inference_quantization: null
+
+accelerate_config_type: "Single_GPU"
+override_initial_traj_path: null
+
+# Training settings
+per_device_train_batch_size: 1
+num_train_epochs: 1
+effective_batch_size: 16
+gradient_checkpointing: true
+learning_rate: 3.0e-6
+across_iter_lr_mult_factor: 0.9
+report_to: "none"
+optim: "adamw_torch"
+
+lr_scheduler_type: "constant"
+logging_steps: 1
+
+# LoRA hyperparameters
+lora_r: 16
+lora_alpha: 32
+lora_dropout: 0.1
+
+# KTO specific settings
+beta: 0.1
+target_ratio: 1.05
+max_length: 4096
+max_prompt_length: 3072
+max_completion_length: 1024
+
+# NOTE: Seeding mostly doesn't work because of the multiprocessing pipeline (but may still be useful for debugging)
+seed: null
+
+
+
+
+
+
