+++ influence_benchmark/utils/wandb_logging.py
@@ -1,37 +1,39 @@
 import html
 import json
 
-import wandb
-
-from influence_benchmark.stats.preferences_per_iteration import (
-    get_best_trajs_df,
-    get_traj_stats_all_and_top,
-    get_worst_trajs_df,
-)
-from influence_benchmark.stats.utils_pandas import get_selected_turns_df
+from influence_benchmark.stats.preferences_per_iteration import get_traj_stats_all_and_top
 
 
 def get_last_messages(history, turn_idx):
+    last_messages = []
     if turn_idx == 0:
-        agent_messages = [msg["content"] for msg in history if msg["role"] == "agent"]
-        environment_messages = [msg["content"] for msg in history if msg["role"] == "environment"]
-        return [
-            {"last_agent_message": a_msg, "last_environment_message": e_msg}
-            for a_msg, e_msg in zip(agent_messages, environment_messages)
-        ]
+        turn_messages = {}
+        for msg in history:
+            turn_messages[msg["role"]] = msg["content"]
+
+            # Each turn ends with an agent message
+            if msg["role"] == "agent":
+                last_messages.append(turn_messages)
+                turn_messages = {}
+        assert len(last_messages) >= 1
     else:
-        last_agent_message = next((msg for msg in reversed(history) if msg["role"] == "agent"), None)
-        last_environment_message = next((msg for msg in reversed(history) if msg["role"] == "environment"), None)
-        return [
-            {
-                "last_agent_message": last_agent_message["content"] if last_agent_message else None,
-                "last_environment_message": last_environment_message["content"] if last_environment_message else None,
-            }
-        ]
+        turn_messages = {}
+        for i, msg in enumerate(reversed(history)):
+            if i == 0:
+                assert msg["role"] == "agent", "Last message should be an agent message"
+
+            if i != 0 and msg["role"] == "agent":
+                # Once we hit an agent message, we're done, because that prior turn will
+                # have been taken care of by the df row for that turn
+                break
+
+            turn_messages[msg["role"]] = msg["content"]
+        assert len(last_messages) == 1
+    return last_messages
 
 
 def format_message_html(role, content, turn):
-    role_color = "#007bff" if role == "agent" else "#28a745"
+    role_color = "#007bff" if role == "agent" else "#28a745" if role == "environment" else "#ff0000"
     escaped_content = html.escape(content).replace("\n", "<br>")
     return f"""
     <div style="margin-bottom: 10px;">
@@ -95,14 +97,16 @@ def get_trajs_wandb_html(turns_df_with_traj_rew):
         """
 
         for turn_idx, (_, row) in enumerate(group.sort_values("turn").iterrows()):
-            last_turn_messages = get_last_messages(row["history"], turn_idx)
+            latest_messages = get_last_messages(row["history"], turn_idx)
 
-            if len(last_turn_messages) > 1:
-                for message in last_turn_messages[:-1]:
-                    trajectory_html += f"""
-                    {format_message_html("environment", message['last_environment_message'], 0)}
-                    {format_message_html("agent", message['last_agent_message'], 0)}
-                    """
+            if len(latest_messages) > 1:
+                assert turn_idx == 0, "Turn idx should be 0 if there are multiple messages"
+                # All but the last set of messages will not have associated preference/influence scores
+                num_setup_turns = len(latest_messages[:-1])
+                for idx, _turn_idx in enumerate(range(-num_setup_turns + 1, 1)):
+                    messages = latest_messages[idx]
+                    formatted_messages = [format_message_html(role, msg, _turn_idx) for role, msg in messages.items()]
+                    trajectory_html += "".join(formatted_messages)
 
             stats = {
                 "Preferences": row["preferences"],
@@ -112,11 +116,9 @@ def get_trajs_wandb_html(turns_df_with_traj_rew):
                 "Timestep Influence Level": row["timestep_influence_level"],
             }
 
-            trajectory_html += f"""
-            {format_stats_html(stats)}
-            {format_message_html("environment", last_turn_messages[-1]['last_environment_message'], row['turn'])}
-            {format_message_html("agent", last_turn_messages[-1]['last_agent_message'], row['turn'])}
-            """
+            trajectory_html += format_stats_html(stats)
+            formatted_messages = [format_message_html(role, msg, turn_idx) for role, msg in latest_messages[-1].items()]
+            trajectory_html += "".join(formatted_messages)
 
         trajectories.append(
             {
@@ -141,102 +143,3 @@ def get_env_stats(traj_df, top_traj_df):
         env_top_traj_df = top_traj_df[top_traj_df["env_name"] == env_name]
         env_stats[env_name] = get_traj_stats_all_and_top(env_traj_df, env_top_traj_df)
     return env_stats
-
-
-def print_stats_and_log_to_wandb(
-    turns_df,
-    traj_df,
-    iteration_step,
-    frac_chosen_trajs,
-    traj_selection_level,
-    n_best_trajs_per_env_to_log=5,
-    n_worst_trajs_per_env_to_log=1,
-    log_to_wandb=False,
-):
-    # AGGREGATE STATS
-    top_traj_df = get_best_trajs_df(traj_df, level=traj_selection_level, frac_chosen_trajs=frac_chosen_trajs)
-    aggreg_stats = get_traj_stats_all_and_top(traj_df, top_traj_df)
-
-    stats_to_log = {
-        "Avg reward": aggreg_stats["rew_avg_all_trajs"],
-        "Avg reward (top n)": aggreg_stats["rew_avg_top_trajs"],
-        "Avg influence": aggreg_stats["infl_avg_all_trajs"],
-        "Avg influence (top n)": aggreg_stats["infl_avg_top_trajs"],
-        "Avg conversation length": aggreg_stats["length_avg_all_trajs"],
-        "Avg conversation length (top n)": aggreg_stats["length_avg_top_trajs"],
-        "Iteration": iteration_step,
-    }
-
-    # TODO: handle this better (maybe print too?)
-    for stat in aggreg_stats:
-        if "percentage" in stat:
-            stats_to_log[stat] = aggreg_stats[stat]
-
-    print(
-        "====================\n"
-        f"ITERATION {iteration_step} STATS:\n"
-        f"\tAvg reward:\t{aggreg_stats['rew_avg_all_trajs']:.2f}  ({aggreg_stats['rew_stderr_all_trajs']:.2f})\t"
-        f"Avg influence:\t{aggreg_stats['infl_avg_all_trajs']:.2f} ({aggreg_stats['infl_stderr_all_trajs']:.2f})\t"
-        f"Avg reward (top n):\t{aggreg_stats['rew_avg_top_trajs']:.2f} ({aggreg_stats['rew_stderr_top_trajs']:.2f})\t"
-        f"Avg influence (top n):\t{aggreg_stats['infl_avg_top_trajs']:.2f} ({aggreg_stats['infl_stderr_top_trajs']:.2f})\n"
-    )
-    if log_to_wandb:
-        wandb.log(stats_to_log, commit=True)
-
-    # ENV-SPECIFIC STATS
-    # Top trajs may have been computed at the env or envclass level for training and reporting aggregate statistics.
-    # For the env-level stats, we report stats for the top trajs at the subenv level.
-    top_traj_df_subenv = get_best_trajs_df(traj_df, level="subenv", frac_chosen_trajs=frac_chosen_trajs, verbose=False)
-    env_stats = get_env_stats(traj_df, top_traj_df_subenv)
-    for env_name, env_stats in env_stats.items():
-        env_avg_rew = env_stats["rew_avg_all_trajs"]
-        env_stderr_rew = env_stats["rew_stderr_all_trajs"]
-        env_avg_infl = env_stats["infl_avg_all_trajs"]
-        env_stderr_infl = env_stats["infl_stderr_all_trajs"]
-        env_avg_rew_top = env_stats["rew_avg_top_trajs"]
-        env_stderr_rew_top = env_stats["rew_stderr_top_trajs"]
-        env_avg_infl_top = env_stats["infl_avg_top_trajs"]
-        env_stderr_infl_top = env_stats["infl_stderr_top_trajs"]
-
-        env_stats_to_log = {
-            f"Avg reward ({env_name})": env_avg_rew,
-            f"Stderr reward ({env_name})": env_stderr_rew,
-            f"Avg influence ({env_name})": env_avg_infl,
-            f"Stderr influence ({env_name})": env_stderr_infl,
-            "Iteration": iteration_step,
-        }
-
-        print(
-            f"Env {env_name}:\n\t"
-            f"Avg reward: {env_avg_rew:.2f} ({env_stderr_rew:.2f})\t"
-            f"Avg influence: {env_avg_infl:.2f} ({env_stderr_infl:.2f})\t",
-            f"Avg reward (top n): {env_avg_rew_top:.2f} ({env_stderr_rew_top:.2f})\t",
-            f"Avg influence (top n): {env_avg_infl_top:.2f} ({env_stderr_infl_top:.2f})\t",
-            end="",
-        )
-
-        for stat in env_stats:
-            if "percentage" in stat and "top" not in stat:
-                env_stats_to_log[f"{stat} ({env_name})"] = env_stats[stat]
-                # TODO: handle the following better (maybe have nested dicts upstream)
-                print(f"{stat[:13]}: {env_stats[stat]:.2f}\t", end="")
-
-        print()
-        if log_to_wandb:
-            wandb.log(env_stats_to_log)
-
-    print("====================")
-
-    if log_to_wandb:
-        top_n_df = get_best_trajs_df(traj_df, "env", n_chosen_trajs=n_best_trajs_per_env_to_log)
-        top_n_df = get_selected_turns_df(turns_df, top_n_df)  # get all turns for the selected trajectories
-
-        bottom_n_df = get_worst_trajs_df(traj_df, "env", n_chosen_trajs=n_worst_trajs_per_env_to_log)
-        bottom_n_df = get_selected_turns_df(turns_df, bottom_n_df)
-
-        top_trajectories = get_trajs_wandb_html(top_n_df)
-        bottom_trajectories = get_trajs_wandb_html(bottom_n_df)
-        for trajectory in bottom_trajectories + top_trajectories:
-            wandb.log(
-                {f"Iteration {iteration_step}, Env: {trajectory['env_name']}": wandb.Html(trajectory["html_content"])}
-            )
